week9-api-deployment/requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
arxiv==2.3.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
azure-ai-agents==1.2.0b6
azure-ai-inference==1.0.0b9
azure-ai-projects==1.1.0b4
azure-common==1.1.28
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.14.2
azure-identity==1.25.1
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
babel==2.17.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.17
chromadb==1.3.5
click==8.3.1
cohere==5.20.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
deepl==1.25.0
defusedxml==0.7.1
deprecated==1.3.1
dill==0.4.0
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docx2txt==0.9
durationpy==0.10
elastic-transport==8.17.1
elasticsearch==8.19.2
emoji==2.15.0
executing==2.2.1
faiss-cpu==1.7.4
fastapi==0.128.0
fastavro==1.12.1
fastjsonschema==2.21.2
feedparser==6.0.12
filelock==3.20.0
filetype==1.2.0
flashrank==0.2.10
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
geojson==3.2.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==1.34.1
google-auth==2.43.0
google-search-results==2.4.2
googleapis-common-protos==1.72.0
grandalf==0.8
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.48.2
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
instructor==1.12.0
ipykernel==6.31.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jpype1==1.6.0
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwipiepy==0.22.0
kiwipiepy-model==0.22.0
kiwisolver==1.4.9
konlpy==0.6.0
kubernetes==33.1.0
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-azure-ai==0.1.2
langchain-azure-ai==0.1.8
langchain-chroma==0.2.2
langchain-chroma==0.2.6
langchain-cohere==0.4.6
langchain-community==0.3.20
langchain-core==0.3.78
langchain-elasticsearch==0.4.0
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-milvus==0.2.1
langchain-openai==0.3.9
langchain-text-splitters==0.3.11
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.6.11
langgraph-checkpoint==3.0.1
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.3.45
lark==1.3.1
llama-cloud==0.1.44
llama-cloud-services==0.6.81
llama-index-core==0.14.8
llama-index-instrumentation==0.4.2
llama-index-readers-file==0.5.5
llama-index-workflows==2.11.4
llama-parse==0.6.81
lxml==5.4.0
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
milvus-lite==2.5.1
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
multiprocess==0.70.18
mypy==1.18.2
mypy-extensions==1.1.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
nodeenv==1.9.1
notebook==7.5.0
notebook-shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.11.0.86
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
posthog==5.4.0
posthog==7.0.1
pre-commit==4.5.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.1.3
ptyprocess==0.7.0
pulsar-client==3.8.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydeck==0.9.1
pygments==2.19.2
pyjwt==2.10.1
pymilvus==2.5.17
pymongo==4.15.4
pyowm==3.5.0
pyparsing==3.2.5
pypdf==6.4.0
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-json-logger==4.0.0
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pywinpty==3.0.2
pyyaml==6.0.3
pyzmq==27.1.0
ragas==0.3.9
rapidfuzz==3.14.3
rapidocr-onnxruntime==1.2.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rouge-score==0.1.2
rpds-py==0.29.0
rsa==4.9.1
scikit-learn==1.7.2
scikit-network==0.33.5
scipy==1.16.3
seaborn==0.13.2
send2trash==1.8.3
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
streamlit==1.51.0
striprtf==0.0.26
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-pyyaml==6.0.12.20250915
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
unstructured==0.18.20
unstructured-client==0.42.4
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
wikipedia==1.4.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0